requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions

joblib==1.3.2
pandas==1.5.3
libpysal==4.9.2
xarray==2023.1.0
pyarrow==11.0.0
scikit-learn==1.3.1
pooch
